refactoring for motif tables and panel
