+++ client/client/app/components/ngbMotifsPanel/ngbMotifsDialog.scss
@@ -3,18 +3,18 @@
   margin-bottom: 0px;
   padding: 0px 20px;
 }
-
+  
 .motifs-dialog-label {
   padding-left: 2px;
   font-weight: 600;
   margin-bottom: 5px;
 }
-
+  
 .motifs-dialog-input {
   max-width: 200px;
   margin-bottom: 0px;
 }
-
+  
 .motifs-dialog-checkbox {
   margin-bottom: 16px;
 
@@ -27,7 +27,8 @@
     margin-bottom: 0.5em;
   }
 }
-
+  
 .motifs-dialog-checkbox-warning {
   font-style: italic;
 }
+
