+++ client/client/app/components/ngbMotifsPanel/ngbMotifsPanel.service.js
@@ -1,60 +1,31 @@
+const PAGE_SIZE = 100;
+const ROW_HEIGHT = 35;
 const REFERENCE = 'REFERENCE';
 const CHROMOSOME = 'CHROMOSOME';
-const POSITIVE = 'POSITIVE';
-const NEGATIVE = 'NEGATIVE';
+const WHOLE_GENOME = 'WHOLE_GENOME';
 const POSITIVE_STRAND = '+';
 const NEGATIVE_STRAND = '-';
+const POSITIVE = 'POSITIVE';
+const NEGATIVE = 'NEGATIVE';
 
 export default class ngbMotifsPanelService {
+
+    searchMotifResults = null;
+    _searchMotifsParams = [];
+    motifsResultsTitle = null;
+    _searchStopOn = {};
+    _currentParams = {};
+    searchRequestsHistory = [];
+    requestNumber = 0;
 
     _isSearchInProgress = false;
     _isSearchFailure = false;
     _errorMessageList = null;
-    _isShowParamsTable = true;
-
-    requestNumber = 0;
-
-    _searchMotifsParams = [];
-    _searchMotifResults = [];
-    searchStopOn = {};
-
-    get referenceType () {
-        return REFERENCE;
-    }
-
-    get chromosomeType () {
-        return CHROMOSOME;
-    }
-
-    get positive () {
-        return POSITIVE;
-    }
-
-    get negative () {
-        return NEGATIVE;
-    }
-
-    get positiveStrand () {
-        return POSITIVE_STRAND;
-    }
-
-    get negativeStrand () {
-        return NEGATIVE_STRAND;
-    }
-
-    static instance (appLayout, dispatcher, projectContext, motifsDataService) {
-        return new ngbMotifsPanelService(appLayout, dispatcher, projectContext, motifsDataService);
-    }
-
-    constructor(appLayout, dispatcher, projectContext, motifsDataService) {
-        Object.assign(this, {appLayout, dispatcher, projectContext, motifsDataService});
-        this.dispatcher.on('reference:change', ::this.panelCloseMotifsPanel);
-    }
+    _isShowParamsTable = false;
 
     get isSearchInProgress () {
         return this._isSearchInProgress;
     }
-
     set isSearchInProgress (value) {
         this._isSearchInProgress = value;
     }
@@ -62,7 +33,6 @@ export default class ngbMotifsPanelService {
     get isSearchFailure () {
         return this._isSearchFailure;
     }
-
     set isSearchFailure (value) {
         this._isSearchFailure = value;
     }
@@ -70,7 +40,6 @@ export default class ngbMotifsPanelService {
     get errorMessageList () {
         return this._errorMessageList;
     }
-
     set errorMessageList (error) {
         this._errorMessageList = error;
     }
@@ -78,7 +47,6 @@ export default class ngbMotifsPanelService {
     get isShowParamsTable () {
         return this._isShowParamsTable;
     }
-
     set isShowParamsTable (value) {
         this._isShowParamsTable = value;
     }
@@ -87,23 +55,181 @@ export default class ngbMotifsPanelService {
         return [...this._searchMotifsParams].reverse();
     }
 
-    set searchMotifsParams (params) {
+    get searchStopOn () {
+        return this._searchStopOn;
+    }
+    set searchStopOn (value) {
+        this._searchStopOn = value;
+    }
+
+    get currentParams () {
+        return this._currentParams;
+    }
+
+    get pageSize () {
+        return PAGE_SIZE;
+    }
+    get rowHeight () {
+        return ROW_HEIGHT;
+    }
+    get referenceType () {
+        return REFERENCE;
+    }
+    get chromosomeType () {
+        return CHROMOSOME;
+    }
+    get wholeGenomeType () {
+        return WHOLE_GENOME;
+    }
+    get positiveStrand () {
+        return POSITIVE_STRAND;
+    }
+    get negativeStrand () {
+        return NEGATIVE_STRAND;
+    }
+    get positive () {
+        return POSITIVE;
+    }
+    get negative () {
+        return NEGATIVE;
+    }
+
+    static instance (
+        appLayout,
+        dispatcher,
+        projectContext,
+        motifsDataService
+    ) {
+        return new ngbMotifsPanelService(
+            appLayout,
+            dispatcher,
+            projectContext,
+            motifsDataService
+        );
+    }
+
+    constructor(
+        appLayout,
+        dispatcher,
+        projectContext,
+        motifsDataService
+    ) {
+        Object.assign(this, {
+            appLayout,
+            dispatcher,
+            projectContext,
+            motifsDataService
+        });
+        this.dispatcher.on('reference:change', this.panelCloseMotifsPanel.bind(this));
+    }
+
+    async searchMotif (params) {
+        this.requestNumber++;
+        this.isSearchInProgress = true;
+        this.isShowParamsTable = false;
+        this.motifsResultsTitle = params.name || params.motif;
+        const request = this.setSearchMotifRequest(params);
+        await this.showSearchMotifResults(request);
+        this.searchRequestsHistory.push(request);
+    }
+
+    panelAddMotifsPanel (params) {
+        this.setMotifsPanel(params);
+        const layoutChange = this.appLayout.Panels.motifs;
+        layoutChange.displayed = true;
+        this.dispatcher.emitSimpleEvent('layout:item:change', {layoutChange});
+    }
+
+    panelCloseMotifsPanel () {
+        this.resetData();
+        const layoutChange = this.appLayout.Panels.motifs;
+        layoutChange.displayed = false;
+        this.dispatcher.emitSimpleEvent('layout:item:change', {layoutChange});
+    }
+
+    async setMotifsPanel (params) {
+        const searchParams = this.setSearchMotifsParams(params);
+        await this.searchMotif(searchParams);
+        const firstHit = (this.searchMotifResults || [])[0];
+        if (firstHit) {
+            this.dispatcher.emitSimpleEvent('motifs:add:tracks', firstHit);
+        }
+    }
+
+    setSearchMotifsParams (params) {
         const currentChromosomeId = this.projectContext.currentChromosome.id;
-        this._searchMotifsParams.push({
+        const searchParams = {
             currentChromosomeId,
             name: params.title,
             motif: params.pattern,
             'search type': params.inReference ?
                 this.referenceType : this.chromosomeType,
+        };
+        this._searchMotifsParams.push(searchParams);
+        return searchParams;
+    }
+
+    setSearchMotifRequest(params) {
+        const referenceType = params['search type'] === this.referenceType;
+        const currentParams = {
+            referenceId: this.projectContext.reference.id,
+            motif: params.motif,
+            searchType: referenceType ?
+                this.wholeGenomeType : this.chromosomeType,
+            pageSize: this.pageSize,
+        };
+        this.setCurrentParams(currentParams);
+        const chromosomeId = params.currentChromosomeId || this.projectContext.currentChromosome.id;
+        return referenceType ?
+            {...currentParams} : {chromosomeId, ...currentParams};
+    }
+
+    setCurrentParams (params) {
+        if (JSON.stringify(params) === '{}') {
+            this._currentParams = {};
+        } else {
+            const name = this.motifsResultsTitle;
+            this._currentParams = {
+                name,
+                ...params
+            };
+        }
+    }
+
+    async showSearchMotifResults(request) {
+        await this.getSearchMotifsResults(request)
+            .then(success => {
+                const result = success ? this.searchMotifResults : [];
+                return this.dispatcher.emitSimpleEvent('motifs:show:results', result);
+            });
+    }
+
+    getSearchMotifsResults(request) {
+        return new Promise((resolve) => {
+            this.motifsDataService.getSearchMotifsResults(request)
+                .then(response => {
+                    this.isSearchInProgress = false;
+                    this.isSearchFailure = false;
+                    this.errorMessageList = null;
+                    this.setSearchMotifResults(response.result);
+                    this.searchStopOn = {
+                        startPosition: response.position !== undefined ?
+                            response.position : null,
+                        chromosomeId: response.chromosomeId || null
+                    };
+                    resolve(true);
+                })
+                .catch((error) => {
+                    this.isSearchInProgress = false;
+                    this.isSearchFailure = true;
+                    this.errorMessageList = [error.message];
+                    resolve(false);
+                });
         });
     }
 
-    get searchMotifResults () {
-        return this._searchMotifResults;
-    }
-
-    set searchMotifResults (result) {
-        this._searchMotifResults = [...(result || []).map(item => {
+    setSearchMotifResults (result) {
+        this.searchMotifResults = [...(result || []).map(item => {
             const strand = this.setStrand(item.strand);
             return {
                 reference: this.projectContext.reference.id,
@@ -133,52 +259,21 @@ export default class ngbMotifsPanelService {
         }
     }
 
-    panelAddMotifsPanel () {
-        const layoutChange = this.appLayout.Panels.motifs;
-        layoutChange.displayed = true;
-        this.dispatcher.emitSimpleEvent('layout:item:change', {layoutChange});
+    resetResultsData () {
+        this._currentParams = {};
+        this.motifsResultsTitle = '';
+        this.searchRequestsHistory = [];
+        this.searchMotifResults = [];
     }
 
-    panelCloseMotifsPanel () {
-        this.resetData();
-        const layoutChange = this.appLayout.Panels.motifs;
-        layoutChange.displayed = false;
-        this.dispatcher.emitSimpleEvent('layout:item:change', {layoutChange});
-    }
-
-    searchMotif (params) {
-        this.requestNumber++;
-        this.searchMotifsParams = params;
-        this.dispatcher.emitSimpleEvent('motifs:search:change');
-        this.panelAddMotifsPanel();
-        this.isShowParamsTable = false;
-        this.dispatcher.emitSimpleEvent('motifs:show:results', this.searchMotifsParams[0]);
-    }
-
-    searchMotifRequest (request) {
-        this.isSearchInProgress = true;
-        return new Promise((resolve) => {
-            this.motifsDataService.getSearchMotifsResults(request)
-                .then(response => {
-                    this.isSearchInProgress = false;
-                    this.isSearchFailure = false;
-                    this.searchMotifResults = response.result;
-                    this.searchStopOn = {
-                        startPosition: response.position !== undefined ? response.position : null,
-                        chromosomeId: response.chromosomeId || null
-                    };
-                    resolve(true);
-                })
-                .catch((error) => {
-                    this.isSearchInProgress = false;
-                    this.isSearchFailure = true;
-                    this.errorMessageList = [error.message];
-                    resolve(false);
-                });
-        });
+    backToParamsTable () {
+        this.resetResultsData();
+        this.isShowParamsTable = true;
+        this.dispatcher.emitSimpleEvent('motifs:show:params');
     }
 
     resetData () {
+        this.resetResultsData();
         this._searchMotifsParams = [];
         this.searchStopOn = {};
     }
