+++ client/client/app/components/ngbMotifsPanel/ngbMotifsParamsTable/ngbMotifsParamsTable.component.js
@@ -1,0 +1,6 @@
+import controller from './ngbMotifsParamsTable.controller';
+
+export default  {
+    controller: controller.UID,
+    template: require('./ngbMotifsParamsTable.tpl.html'),
+};
