+++ client/client/app/components/ngbMotifsPanel/ngbMotifsParamsTable/ngbMotifsParamsTable.tpl.html
@@ -1,0 +1,11 @@
+<div class="ui-grid-container">
+    <div ui-grid="$ctrl.gridOptions"
+         ui-grid-resize-columns
+         ui-grid-move-columns
+         ui-grid-auto-resize
+         ui-grid-selection
+         ui-grid-save-state
+         ui-grid-pinning
+         class="motifs-params-table">
+    </div>
+</div>
