+++ client/client/app/components/ngbMotifsPanel/ngbMotifsResultsTable/ngbMotifsResultsTable.component.js
@@ -1,0 +1,9 @@
+import controller from './ngbMotifsResultsTable.controller';
+
+export default  {
+    controller: controller.UID,
+    template: require('./ngbMotifsResultsTable.tpl.html'),
+    bindings: {
+        loading: '<'
+    }
+};
