+++ client/client/app/components/ngbMotifsPanel/ngbMotifsResultsTable/ngbMotifsResultsTable.scss
@@ -1,0 +1,33 @@
+.motifs-results-title {
+    align-self: center;
+}
+  
+.motifs-back-to-params-button {
+    min-width: 44px;
+}
+  
+.motifs-results-table {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+
+    .ui-grid-viewport {
+        height: 100% !important;
+    }
+
+    .ui-grid-header-cell-label {
+        text-transform: capitalize;
+    }
+}
+
+.motifs-empty-results-container {
+    font-size: 12pt;
+    width: 100%;
+    text-align: center;
+    z-index: 1000;
+}
+  
+.motifs-results-no-records-warning {
+    color: #999;
+}
+
