+++ client/client/app/components/ngbMotifsPanel/ngbMotifsResultsTable/ngbMotifsResultsTable.tpl.html
@@ -1,13 +1,15 @@
 <div class="ui-grid-container">
-    <div ng-if="!$ctrl.isShowParamsTable" flex="100" layout="row">
+    <div flex="100" layout="row">
         <md-button
             aria-label="back"
-            class="motifs-back-button"
+            class="motifs-back-to-params-button"
             ng-click="$ctrl.backToParamsTable()"
             type="button">
             <ng-md-icon icon="keyboard_arrow_left" size="24"></ng-md-icon>
         </md-button>
-        <span flex="50" class="motifs-search-title">{{$ctrl.motifsSearchTitle}}</span>
+        <span flex="50" class="motifs-results-title">
+            {{$ctrl.motifsResultsTitle}}
+        </span>
     </div>
     <div ui-grid="$ctrl.gridOptions"
          ui-grid-resize-columns
@@ -18,12 +20,12 @@
          ui-grid-save-state
          ui-grid-pinning
          ng-hide="$ctrl.emptyResults"
-         class="motifs-search-table">
+         class="motifs-results-table">
     </div>
-    <div class="motif-results-empty-result-container"
+    <div class="motifs-empty-results-container"
          ng-if="$ctrl.emptyResults"
     >
-        <md-content class="motif-results-no-records-warning"
+        <md-content class="motifs-results-no-records-warning"
                     flex>
             <span>No records found.</span>
         </md-content>
