--- client/client/app/components/ngbMotifsPanel/ngbMotifsTable/ngbMotifsTable.component.js
@@ -1,9 +1,0 @@
-import controller from './ngbMotifsTable.controller';
-
-export default  {
-    controller: controller.UID,
-    template: require('./ngbMotifsTable.tpl.html'),
-    bindings: {
-        loading: '<'
-    }
-};
