--- client/client/app/components/ngbMotifsPanel/ngbMotifsTable/ngbMotifsTable.scss
@@ -1,31 +1,0 @@
-.motifs-search-title {
-  align-self: center;
-}
-
-.motifs-back-button {
-  min-width: 44px;
-}
-
-.motifs-search-table {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-    .ui-grid-viewport {
-      height: 100% !important;
-    }
-
-    .ui-grid-header-cell-label {
-      text-transform: capitalize;
-    }
-}
-
-.motif-results-empty-result-container {
-  font-size: 12pt;
-  width: 100%;
-  text-align: center;
-  z-index: 1000;
-}
-
-.motif-results-no-records-warning {
-  color: #999;
-}
